requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.5.2
chardet==5.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
eventlet==0.40.1
fastapi==0.115.12
flower==2.0.1
greenlet==3.2.3
h11==0.14.0
humanize==4.12.3
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
pydantic==2.11.3
pydantic_core==2.33.1
pypdf==5.4.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
redis==6.1.0
reportlab==4.4.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tornado==6.5.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
